+++ module-apps/application-desktop/ApplicationDesktop.cpp
@@ -22,9 +22,11 @@
 #include <cassert>
 namespace app
 {
+    static const unsigned int default_screen_pin_size = 4;
+    static const unsigned int default_screen_attempts = 4;
 
     ApplicationDesktop::ApplicationDesktop(std::string name, std::string parent, bool startBackground)
-        : Application(name, parent)
+        : Application(name, parent), lock(this)
     {
         busChannels.push_back(sys::BusChannels::ServiceDBNotifications);
     }
@@ -136,7 +138,7 @@ namespace app
     {
         assert(msg);
         if (msg->request == cellular::State::ST::URCReady) {
-            if (need_sim_select && !screenLocked) {
+            if (need_sim_select && !lock.isLocked()) {
                 sapm::ApplicationManager::messageSwitchApplication(this, app::name_settings, app::sim_select, nullptr);
                 return true;
             }
@@ -149,7 +151,6 @@ namespace app
         if (msg->request == cellular::State::ST::ModemFatalFailure) {
             switchWindow(app::window::name::desktop_reboot);
         }
-
         return false;
     }
 
@@ -204,8 +205,6 @@ namespace app
             return ret;
         }
 
-        screenLocked = true;
-
         reloadSettings();
         requestNotReadNotifications();
         requestNotSeenNotifications();
@@ -251,27 +250,13 @@ namespace app
     {
         using namespace app::window::name;
         windows.insert(std::pair<std::string, gui::AppWindow *>(desktop_main_window, new gui::DesktopMainWindow(this)));
-        windows.insert(std::pair<std::string, gui::AppWindow *>(desktop_pin_lock, new gui::PinLockWindow(this)));
+        windows.insert(std::pair<std::string, gui::AppWindow *>(desktop_pin_lock,
+                                                                new gui::PinLockWindow(this, desktop_pin_lock, lock)));
         windows.insert(std::pair<std::string, gui::AppWindow *>(desktop_menu, new gui::MenuWindow(this)));
         windows.insert(std::pair<std::string, gui::AppWindow *>(desktop_poweroff, new gui::PowerOffWindow(this)));
         windows.insert(std::pair<std::string, gui::AppWindow *>(desktop_reboot, new gui::RebootWindow(this)));
         windows.insert(std::pair<std::string, gui::AppWindow *>(desktop_update, new gui::UpdateWindow(this)));
     }
-
-    bool ApplicationDesktop::getScreenLocked()
-    {
-        return screenLocked;
-    }
-
-    bool ApplicationDesktop::getPinLocked()
-    {
-        return pinLocked;
-    }
-
-    void ApplicationDesktop::setScreenLocked(bool val)
-    {
-        screenLocked = val;
-    };
 
     void ApplicationDesktop::destroyUserInterface()
     {}
@@ -279,9 +264,6 @@ namespace app
     void ApplicationDesktop::reloadSettings()
     {
         settings = DBServiceAPI::SettingsGet(this);
-
-        pinLocked = settings.lockPassHash != 0;
-
         switch (settings.activeSIM) {
         case SettingsRecord::ActiveSim::NONE:
             Store::GSM::get()->selected = Store::GSM::SIM::NONE;
@@ -293,6 +275,32 @@ namespace app
         case SettingsRecord::ActiveSim::SIM2:
             Store::GSM::get()->selected = Store::GSM::SIM::SIM2;
             break;
+        }
+
+        unsigned int pinSize = settings.lockPassHash == 0 ? 0 : default_screen_pin_size;
+        lock.reset(gui::PinLock::LockType::Screen, gui::PinLock::State::EnterPin, default_screen_attempts, pinSize);
+    }
+
+    void ApplicationDesktop::handlePin(const std::vector<unsigned int> &pin)
+    {
+        if (lock.getLockType() == gui::PinLock::LockType::Screen) {
+            if (lock.remainingAttempts > 0) {
+                uint32_t hash = 0;
+                for (auto i : pin) {
+                    hash = 10 * hash + i;
+                }
+                lock.remainingAttempts--;
+                if (hash == settings.lockPassHash) {
+                    lock.remainingAttempts = 4;
+                    lock.state             = gui::PinLock::State::VerifiedPin;
+                }
+                else if (lock.remainingAttempts > 1) {
+                    lock.state = gui::PinLock::State::InvalidPin;
+                }
+                else {
+                    lock.state = gui::PinLock::State::Blocked;
+                }
+            }
         }
     }
 } // namespace app