+++ module-apps/application-desktop/ApplicationDesktop.hpp
@@ -1,6 +1,7 @@
 #pragma once
 
 #include "windows/Names.hpp"
+#include "widgets/PinLock.hpp"
 
 #include <Application.hpp>
 #include <Service/Message.hpp>
@@ -18,9 +19,6 @@ namespace app
     class ApplicationDesktop : public Application
     {
       protected:
-        // determines whether screen should be protected by pin verification
-        bool screenLocked = true;
-        bool pinLocked    = false;
         void reloadSettings();
 
       public:
@@ -43,6 +41,8 @@ namespace app
 
         } notifications;
 
+        gui::PinLock lock;
+
         ApplicationDesktop(std::string name = name_desktop, std::string parent = "", bool startBackground = false);
         virtual ~ApplicationDesktop();
         sys::Message_t DataReceivedHandler(sys::DataMessage *msgl, sys::ResponseMessage *resp) override;
@@ -53,12 +53,8 @@ namespace app
         {
             return sys::ReturnCodes::Success;
         }
-
         void createUserInterface() override;
         void destroyUserInterface() override;
-        bool getScreenLocked();
-        void setScreenLocked(bool val);
-        bool getPinLocked();
         // if there is modem notification and there is no default SIM selected, then we need to select if when unlock is
         // done
         bool handle(db::NotificationMessage *msg);
@@ -76,6 +72,8 @@ namespace app
         bool clearMessagesNotification();
         bool requestNotSeenNotifications();
         bool requestNotReadNotifications();
+
+        void handlePin(const std::vector<unsigned int> &);
     };
 
 } /* namespace app */