+++ module-apps/application-desktop/windows/DesktopMainWindow.hpp
@@ -1,6 +1,5 @@
 #pragma once
 
-#include "../ApplicationDesktop.hpp"
 #include "AppWindow.hpp"
 #include "gui/widgets/BottomBar.hpp"
 #include "gui/widgets/Image.hpp"
@@ -8,6 +7,11 @@
 #include "gui/widgets/Text.hpp"
 #include "gui/widgets/TopBar.hpp"
 #include "gui/widgets/Window.hpp"
+
+namespace app
+{
+    class ApplicationDesktop;
+}
 
 namespace gui
 {
@@ -59,6 +63,10 @@ namespace gui
         // method hides or show widgets and sets bars according to provided state
         void setVisibleState();
         auto fillNotifications(app::ApplicationDesktop *app) -> bool;
+        bool processLongPressEvent(const InputEvent &inputEvent);
+        bool processShortPressEventOnUnlocked(const InputEvent &inputEvent);
+        bool processShortPressEventOnLocked(const InputEvent &inputEvent);
+        app::ApplicationDesktop *getAppDesktop() const;
 
       public:
         DesktopMainWindow(app::Application *app);