+++ module-db/Interface/EventsRecord.cpp
@@ -52,7 +52,6 @@ bool EventsRecordInterface::Add(const EventsRecord &rec)
         return eventsDb->events.addCustom(entry);
     }
     }
-
 }
 
 std::unique_ptr<std::vector<EventsRecord>> EventsRecordInterface::Select(TimePoint filter_from, TimePoint filter_till)